Reorganise AMW_angular resources directory structure
