+++ AMW_angular/io/src/app/apps/app-add/app-add.component.ts
@@ -2,9 +2,9 @@ import { Component, EventEmitter, Input, Output, Signal } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { NgSelectModule } from '@ng-select/ng-select';
 import { Release } from '../../settings/releases/release';
-import { Release as Rel } from '../../resource/release';
+import { Release as Rel } from '../../resources/models/release';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { Resource } from '../../resource/resource';
+import { Resource } from '../../resources/models/resource';
 import { AppCreate } from '../app-create';
 import { ModalHeaderComponent } from '../../shared/modal-header/modal-header.component';
 import { ButtonComponent } from '../../shared/button/button.component';
