+++ AMW_angular/io/src/app/deployment/deployment.component.spec.ts
@@ -4,10 +4,10 @@ import { FormsModule } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { of, Subject } from 'rxjs';
-import { Release } from '../resource/release';
-import { Resource } from '../resource/resource';
-import { ResourceTag } from '../resource/resource-tag';
-import { ResourceService } from '../resource/resource.service';
+import { Release } from '../resources/models/release';
+import { Resource } from '../resources/models/resource';
+import { ResourceTag } from '../resources/models/resource-tag';
+import { ResourceService } from '../resources/services/resource.service';
 import { AppWithVersion } from './app-with-version';
 import { Deployment } from './deployment';
 import { DeploymentParameter } from './deployment-parameter';
