+++ AMW_angular/io/src/app/deployments/deployments.component.ts
@@ -5,7 +5,7 @@ import { ActivatedRoute, Params } from '@angular/router';
 import * as _ from 'lodash';
 import * as datefns from 'date-fns';
 import { Subscription, timer } from 'rxjs';
-import { ResourceService } from '../resource/resource.service';
+import { ResourceService } from '../resources/services/resource.service';
 import { DeploymentFilter } from '../deployment/deployment-filter';
 import { DeploymentFilterType } from '../deployment/deployment-filter-type';
 import { ComparatorFilterOption } from '../deployment/comparator-filter-option';
