+++ AMW_angular/io/src/app/resources/models/property.ts
