+++ AMW_angular/io/src/app/resources/models/resource-template.ts
