+++ AMW_angular/io/src/app/resources/models/resource-type-request.ts
