+++ AMW_angular/io/src/app/resources/models/resource.ts
