+++ AMW_angular/io/src/app/resources/resource-edit/resource-functions/resource-functions-list.component.spec.ts
@@ -3,7 +3,7 @@ import { InputSignal, signal } from '@angular/core';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ResourceFunctionsListComponent } from './resource-functions-list.component';
-import { Resource } from '../../../resource/resource';
+import { Resource } from '../../models/resource';
 
 describe('ResourceFunctionsComponent', () => {
   let component: ResourceFunctionsListComponent;
