+++ AMW_angular/io/src/app/resources/resource-type-delete/resource-type-delete.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, inject, Input, Output } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { ResourceType } from '../../resource/resource-type';
+import { ResourceType } from '../models/resource-type';
 import { ModalHeaderComponent } from '../../shared/modal-header/modal-header.component';
 import { ButtonComponent } from '../../shared/button/button.component';
