+++ AMW_angular/io/src/app/resources/resources-page.component.scss
