+++ AMW_angular/io/src/app/resources/resources-page.component.spec.ts
@@ -11,7 +11,9 @@ describe('ResourcesPageComponent', () => {
 
   const mockRoute = {
     snapshot: {},
-    queryParams: of({}),
+    queryParamMap: of({
+      get: () => null,
+    }),
   };
 
   beforeEach(async () => {
