+++ AMW_angular/io/src/app/settings/permission/permission.component.spec.ts
@@ -15,8 +15,8 @@ import { RestrictionListComponent } from './restriction-list.component';
 import { Restriction } from './restriction';
 import { Tag } from './tag';
 import { EnvironmentService } from '../../deployment/environment.service';
-import { ResourceService } from '../../resource/resource.service';
-import { ResourceTypesService } from '../../resource/resource-types.service';
+import { ResourceService } from '../../resources/services/resource.service';
+import { ResourceTypesService } from '../../resources/services/resource-types.service';
 import { Environment } from 'src/app/deployment/environment';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
