+++ AMW_angular/io/src/app/settings/permission/restriction-add.component.ts
@@ -5,8 +5,8 @@ import { RestrictionsCreation } from './restrictions-creation';
 import { Permission } from './permission';
 import * as _ from 'lodash';
 import { Environment } from 'src/app/deployment/environment';
-import { Resource } from 'src/app/resource/resource';
-import { ResourceType } from 'src/app/resource/resource-type';
+import { Resource } from 'src/app/resources/models/resource';
+import { ResourceType } from 'src/app/resources/models/resource-type';
 import { IconComponent } from '../../shared/icon/icon.component';
 import { FormsModule } from '@angular/forms';
 import { NgSelectModule } from '@ng-select/ng-select';
